rename tests files and folders
