+++ src/app/uicomponents/GameOverModalWindowContent/Errors.js
