+++ src/app/uicomponents/GameOverModalWindowContent/GameOverModalWindowContent.css
