+++ src/app/uicomponents/GameOverModalWindowContent/GameOverModalWindowContent.js
