+++ src/app/uicomponents/PeopleImageToRecognize/PeopleImageToRecognize.css
