+++ src/app/uicomponents/PeopleImageToRecognize/PeopleImageToRecognize.js
